Please read everything in the message,
What I've been doing on the reconfiguration so far
It's not completely done, but for now, only moving blocks (yeah! ;p) is
the leftover part to implement.

Also take into account that the pathfinding algorithm has to run first,
to make this part to work. However, I can imagine that Rico will fuck
with us and place all Robot blocks in the same region as the Target
blocks. In that case we need something else.